requirements.txt
pandas>=2.0.0
numpy>=1.24.0
# spacy>=3.6.0
nltk>=3.8.1
textblob>=0.17.1
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0
python-dotenv>=1.0.0
jupyter>=1.0.0
# PyPDF2>=3.0.0
# pdfplumber>=0.10.0
camelot-py[cv]>=0.11.0
opencv-python>=4.8.0
ghostscript>=0.7.0
